pyproject.toml
[project]
name = "lazy-ta"
version = "0.1.0"
description = "懶惰的 TA 不想改上機考 🙃"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.20.0",
]

[dependency-groups]
dev = [
    {include-group = "test"},
    "pre-commit>=4.3.0",
    "ruff>=0.14.2",
]
test = [
    "pytest>=8.4.2",
]


[tool.ruff]
lint.select = [
    "Q",
    "I",
    "F401",
]
lint.exclude = [
    ".git",
    ".ruff_cache",
    "logs",
    "media",
    "static",
    "venv",
    ".venv",
]
lint.ignore-init-module-imports = true
line-length = 100

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "double"
docstring-quotes = "double"
